+++ package.json
@@ -39,6 +39,8 @@
     "multer-s3": "^3.0.1",
     "pg": "^8.11.3",
     "reflect-metadata": "^0.2.1",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.0",
     "ts-node": "^10.9.2",
     "typeorm": "^0.3.20"
   },
@@ -50,6 +52,8 @@
     "@types/multer": "^1.4.11",
     "@types/multer-s3": "^3.0.3",
     "@types/node": "^20.11.20",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.6",
     "cross-var": "^1.1.0",
     "jest": "^29.7.0",
     "nodemon": "^3.1.0",
